Make more mobile friendly
Fortunately the main responsive behaviour is already handled by using the
proper bootstrap classes.
I did decide to make the navbar collapsable since it would be the best
experience on a small screen. This require me to add jquery and some bootstrap
javascript to the page and tweak the navbar styles a bit though it did lead me
to realize I had some unneccesary styles on the navbar.